refactor(auth): Move auth elements to its own module
In order to move forward with the auth database table, we are now encapsulating everything inside
the auth module.

connects to #21
