+++ tomaco/tests/__init__.py
@@ -1,0 +1,6 @@
+from flask import url_for
+
+
+class BaseTest:
+    def get(self, client):
+        return client.get(url_for(self.url_name))
